Add Custom Fact for MobileConfig Profiles
This commit adds a really rough fact to gather installed mobile
config profiles and assign them to a new fact, macosx_mobileconfig_profiles.

This fact currently just calls out to the profiles, grep then awk
and should be refactored for in-Ruby string manipulation but this
satisfies my needs for a 0.0.2 Forge release.