Refactor: Use Native Type & Fact
This commit completely scraps the Google python script/service way
of managing mobileconfig profiles.

Instead, the profile_manager type is created which simply works
with the profiles command to create, destroy and check on profiles.
A stupid simple fact is created to show all included profiles.

The type could be improved to do parsing in Ruby and to support user
versus computer or domain profiles and the fact could be improved
to use system_profiler and retrieve either structured data (facter
release willing) or return a new fact for each installed profile.